Update script and configs

+++ install-nix.sh
@@ -4,20 +4,28 @@ if [ $(id -u) != 0 ]; then
    sudo -v
 fi
 
-# Fix locale errors
-export LC_ALL="en_US.UTF-8"
-export LANGUAGE="en_US.UTF-8"
-
 # Windows <-> Linux clipboard support
-mkdir ~/lemonade_tmp
-wget -P ~/lemonade_tmp/ -q https://github.com/pocke/lemonade/releases/download/v1.1.1/lemonade_linux_amd64.tar.gz
-tar xzf ~/lemonade_tmp/lemonade_linux_amd64.tar.gz -C ~/lemonade_tmp/
-sudo cp ~/lemonade_tmp/lemonade /usr/local/bin
-rm -rf ~/lemonade_tmp/
+if ! type -a lemonade &> /dev/null; then
+   mkdir ~/lemonade_tmp
+   wget -P ~/lemonade_tmp/ -q https://github.com/pocke/lemonade/releases/download/v1.1.1/lemonade_linux_amd64.tar.gz
+   tar xzf ~/lemonade_tmp/lemonade_linux_amd64.tar.gz -C ~/lemonade_tmp/
+   sudo cp ~/lemonade_tmp/lemonade /usr/local/bin
+   rm -rf ~/lemonade_tmp/
+fi
 
 # Nix package manager
-curl https://nixos.org/nix/install | sh
-source ~/.nix-profile/etc/profile.d/nix.sh
+if ! type -a nix-env &> /dev/null; then
+   sudo mkdir /nix
+   sudo chown -R ${USER}:$(id -gn $USER) /nix
+   curl https://nixos.org/nix/install | sh
+   source ~/.nix-profile/etc/profile.d/nix.sh
+fi
+
+# Fix locale errors
+nix-env -iA nixpkgs.glibcLocales
+export LC_ALL="en_US.UTF-8"
+export LANGUAGE="en_US.UTF-8"
+export LOCALE_ARCHIVE="$(readlink ~/.nix-profile/lib/locale)/locale-archive"
 
 # Git
 nix-env -i git
@@ -26,14 +34,12 @@ nix-env -i git
 nix-env -i tmux
 
 # Neovim + plugins
-nix-env -i neovim python27Packages.neovim python36Packages.neovim
+nix-env -iA nixpkgs.neovim nixpkgs.python27Packages.neovim nixpkgs.python36Packages.neovim
 
-# Zsh + Zgen
+# Zsh + Zplug
 nix-env -i zsh
-git clone https://github.com/tarjoilija/zgen.git "${HOME}/.zgen"
-
-# Set default shell
-chsh -s $(which tmux)
+curl -sL --proto-redir -all,https https://raw.githubusercontent.com/zplug/installer/master/installer.zsh| zsh
+chmod -R 755 ~/.zplug
 
 # Gnu Stow
 nix-env -i stow
@@ -51,15 +57,28 @@ nvim --headless +PlugInstall +qall
 # Redis/Node/Psql
 nix-env -i redis nodejs postgresql
 
-# Ruby/Elixir
-nix-env -i ruby
-nix-env -i elixir
+# Asdf version manager
+git clone https://github.com/asdf-vm/asdf.git "${HOME}/.asdf" --branch v0.3.0
+source ~/.asdf/asdf.sh
 
-# Ruby gems
+# Ruby
+asdf plugin-add ruby https://github.com/asdf-vm/asdf-ruby.git
+asdf install ruby 2.4.0
+asdf global ruby 2.4.0
+
 zsh -c "gem install bundler"
 zsh -c "gem install foreman"
 zsh -c "gem install rails -v 5.0.1"
 
-# Elixir mix
-zsh -c "mix local.hex"
-zsh -c "mix archive.install https://github.com/phoenixframework/archives/raw/master/phx_new.ez"
+# Elixir
+asdf plugin-add erlang https://github.com/asdf-vm/asdf-erlang.git
+asdf plugin-add elixir https://github.com/asdf-vm/asdf-elixir.git
+
+asdf install erlang 20.0
+asdf global erlang 20.0
+
+asdf install elixir 1.5.0
+asdf global elixir 1.5.0
+
+zsh -c "mix local.hex --force"
+zsh -c "mix archive.install https://github.com/phoenixframework/archives/raw/master/phx_new.ez --force"
